Lv2-2

+++ src/main/java/com/example/calculator/Cal1.java
@@ -14,8 +14,8 @@ public class Cal1 {
         calculation = new ArrayList<>();
     }
     //사칙연산을 수행한 후, 결과값을 반환하는 메서드 구현
-    public int calculate(int num1,int num2, char oper){
-        int result;
+    public double calculate(double num1,double num2, char oper){
+        double result;
 
         switch (oper){
             case '+':
